+++ packages/core-js/full/symbol/async-dispose.js
@@ -1,5 +1,4 @@
 'use strict';
-require('../../modules/esnext.symbol.async-dispose');
-var WrappedWellKnownSymbolModule = require('../../internals/well-known-symbol-wrapped');
+var parent = require('../../actual/symbol/async-dispose');
 
-module.exports = WrappedWellKnownSymbolModule.f('asyncDispose');
+module.exports = parent;