+++ packages/core-js/proposals/async-explicit-resource-management.js
@@ -1,4 +1,5 @@
 'use strict';
+// TODO: Remove from `core-js@4`
 // https://github.com/tc39/proposal-async-explicit-resource-management
 require('../modules/esnext.suppressed-error.constructor');
 require('../modules/esnext.async-disposable-stack.constructor');