+++ CHANGELOG.md
@@ -8,6 +8,19 @@
   - Added `/actual/` namespaces entries, unconditional forced replacement changed to feature detection
 - [Iterator Helpers stage 3 proposal](https://github.com/tc39/proposal-iterator-helpers):
   - Avoid creating observable `String` wrapper objects, [proposal-iterator-helpers/281](https://github.com/tc39/proposal-iterator-helpers/pull/281)
+- Async explicit resource management:
+  - Built-ins:
+    - `Symbol.asyncDispose`
+    - `AsyncDisposableStack`
+      - `AsyncDisposableStack.prototype.disposeAsync`
+      - `AsyncDisposableStack.prototype.use`
+      - `AsyncDisposableStack.prototype.adopt`
+      - `AsyncDisposableStack.prototype.defer`
+      - `AsyncDisposableStack.prototype.move`
+      - `AsyncDisposableStack.prototype[@@asyncDispose]`
+    - `AsyncIterator.prototype[@@asyncDispose]`
+  - Moved back into [the initial proposal](https://github.com/tc39/proposal-explicit-resource-management) -> moved to stage 3, [proposal-explicit-resource-management/154](https://github.com/tc39/proposal-explicit-resource-management/pull/154)
+  - Added `/actual/` namespace entries, disabled forced replacement
 - Used strict mode in some missed cases, [#1269](https://github.com/zloirock/core-js/issues/1269)
 - Compat data improvements:
   - [NodeJS 20.4.0 add `Symbol.{ dispose, asyncDispose }`](https://github.com/nodejs/node/pull/48518), but [with incorrect property descriptors](https://github.com/nodejs/node/issues/48699)